feat: api access token

+++ src/pages/api/[[...user]].tsx
@@ -1,5 +1,7 @@
 import { deleteData, retrieveData, updateData } from "@/lib/firebase/service";
+import { console } from "inspector";
 import type { NextApiRequest, NextApiResponse } from "next";
+import  Jwt  from "jsonwebtoken";
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
     if (req.method === "GET") {
@@ -16,23 +18,43 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
             data
         })
     }else if(req.method === "PUT"){
-        const {id, data} = req.body;
-        await updateData('users', id, data, (result: boolean) => {
-            if(result){
-            res.status(200).json({status: true, statusCode: 200, message: "success"});
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        const {user}:any = req.query;
+        const {data} = req.body;
+        const token = req.headers.authorization?.split(' ')[1] || '';
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        Jwt.verify(token, process.env.NEXTAUTH_SECRET || '', async (err: any, decoded: any) => {
+            if(decoded && decoded.role === 'admin'){
+                await updateData('users', user[1], data, (result: boolean) => {
+                if(result){
+                    res.status(200).json({status: true, statusCode: 200, message: "success"});
+                }else{
+                    res.status(400).json({status: false, statusCode: 400, message: "failed"});
+                }
+                });
             }else{
-            res.status(400).json({status: false, statusCode: 400, message: "failed"});
+                res.status(403).json({status: false, statusCode: 403, message: "Access Denied"});
             }
-        });
+        })
+       
         
     }else if(req.method === "DELETE"){
         // eslint-disable-next-line @typescript-eslint/no-explicit-any
         const {user}: any = req.query;
-        await deleteData('users',user[1], (result: boolean) => {
-            if(result){
-                res.status(200).json({status: true, statusCode: 200, message: "success"});
+        const token = req.headers.authorization?.split(' ')[1] || '';
+        console.log(token)
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        Jwt.verify(token, process.env.NEXTAUTH_SECRET || '', async (err: any, decoded: any) => {
+            if(decoded && decoded.role === 'admin'){
+                await deleteData('users',user[1], (result: boolean) => {
+                    if(result){
+                        res.status(200).json({status: true, statusCode: 200, message: "success"});
+                    }else{
+                        res.status(400).json({status: false, statusCode: 400, message: "failed"});
+                    }
+                })
             }else{
-                res.status(400).json({status: false, statusCode: 400, message: "failed"});
+                res.status(403).json({status: false, statusCode: 403, message: "Access Denied"});
             }
         })
     }
